action_creators: move preparePageForSave to page.ts
