+++ superglue/lib/action_creators/index.ts
@@ -1,3 +1,3 @@
 export * from './requests'
-export { saveAndProcessPage } from './page'
+export { saveAndProcessPage, preparePageForSave } from './page'
 export { webVisit, webRemote } from './web'
