+++ superglue/lib/action_creators/requests.ts
@@ -6,7 +6,6 @@ import {
   hasPropsAt,
   propsAtParam,
   removePropsAt,
-  dangerouslyEachIn,
 } from '../utils'
 import {
   beforeFetch,
@@ -16,11 +15,10 @@ import {
   superglueError,
   receiveResponse,
 } from '../actions'
-import { saveAndProcessPage } from './index'
+import { saveAndProcessPage, preparePageForSave } from './index'
 import {
   FetchArgs,
   PageResponse,
-  Page,
   SuperglueState,
   Result,
   ErrorResult,
@@ -30,72 +28,9 @@ import {
   NavigationAction,
   VisitResult,
   BeforeSave,
-  AllFragments,
-  GraftResponse,
-  SaveResponse,
-  JSONMappable,
 } from '../types'
 import { createProxy } from '../utils/proxy'
 import { SuperglueResponseError } from '../utils/request'
-
-export function preparePageForSave<T extends JSONMappable = JSONMappable>(
-  nextPage: GraftResponse<T> | SaveResponse<T>,
-  prevPage: Page<T> | undefined,
-  prevFragments: AllFragments,
-  beforeSave: BeforeSave<T>
-): GraftResponse<T> | SaveResponse<T> {
-  const existingPage = prevPage
-    ? createProxy(
-        prevPage,
-        { current: prevFragments },
-        new Set(),
-        new WeakMap()
-      )
-    : undefined
-
-  nextPage.fragments
-    .slice()
-    .reverse()
-    .forEach((fragment) => {
-      const { path } = fragment
-
-      dangerouslyEachIn(nextPage, path, (child, key, nextKey) => {
-        if (Array.isArray(child) && key && !key.includes('=')) {
-          Object.freeze(child)
-        }
-
-        if (nextKey && !nextKey.includes('=') && parseInt(nextKey)) {
-          if (Array.isArray(child)) {
-            Object.defineProperty(child, nextKey, {
-              writable: false,
-              configurable: false,
-            })
-          }
-
-          if (
-            typeof child === 'object' &&
-            child !== null &&
-            !Array.isArray(child)
-          ) {
-            Object.defineProperty(child, nextKey, {
-              writable: false,
-              configurable: false,
-            })
-          }
-        }
-      })
-    })
-
-  return JSON.parse(
-    JSON.stringify(beforeSave(existingPage, nextPage), (_key, value) => {
-      if (value && typeof value === 'object' && value.__id) {
-        return { __id: value.__id }
-      }
-
-      return value
-    })
-  ) as typeof nextPage
-}
 
 /**
  * Handles a thunk rejection using a discriminated `ErrorResult`
